+++ src/main/java/com/centralesupelec/chowchow/user/domain/UserRepository.java
@@ -3,7 +3,7 @@ package com.centralesupelec.chowchow.user.domain;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<UserEntity, String> {
-  UserEntity findById(final Long id);
+  UserEntity findById(final Integer id);
 
   UserEntity findByUsername(final String username);
 }
